improve tax-rate service and mapper

+++ backend/src/main/java/core/hubby/backend/tax/controller/dto/TaxRateRequests.java
@@ -1,5 +1,6 @@
 package core.hubby.backend.tax.controller.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -11,7 +12,7 @@ public record TaxRateRequests(
 		String taxType,
 		List<Component> taxComponents,
 		@NotNull(message = "effectiveRate component cannot be null.")
-		Double effectiveRate,
+		BigDecimal effectiveRate,
 		ApplyToAccounts ApplyToAccounts,
 		Boolean systemDefined,
 		@NotBlank(message = "status component cannot be blank.")
@@ -21,7 +22,7 @@ public record TaxRateRequests(
 			@NotBlank(message = "name component cannot be blank.")
 			String name,
 			@NotNull(message = "rate component cannot be null.")
-			Double rate,
+			BigDecimal rate,
 			Boolean isCompound,
 			Boolean nonRecoverable
 	) {}
